+++ module/dramaticScene/torgeternityCombat.js
@@ -278,7 +278,8 @@ export default class TorgCombat extends Combat {
     await this.drawDramaCard();
 
     // deactivate active defense when the combat round is progressed. End of combat is in the hook above, 'deleteCombat'
-    await this.#deleteActiveDefense();
+    if (game.release.generation < 14)
+      await this.#deleteActiveDefense();
 
     // Perform end-of-faction's turn processing
     return this.nextRoundKeep();
@@ -337,6 +338,18 @@ export default class TorgCombat extends Combat {
   async dramaFlurry(faction) {
     // extra turn
     console.log('Drama Flurry', faction)
+    // Reset the "turnTaken" for all members of this faction (keep the same TURN number in the combat tracker).
+    await this.updateEmbeddedDocuments('Combatant',
+      this.combatants.filter(combatant => this.getCombatantFaction(combatant) === faction)
+        .map((combatant) => ({
+          _id: combatant.id,
+          'system.turnTaken': false,
+          'system.multiAction': null
+        })),
+      { updateAll: true });
+    this.setCardsPlayable(true);
+    this.unsetFlag('torgeternity', FATIGUED_FACTION_FLAG);
+
     return this.#sendDramaChat('flurry', faction);
   }