+++ module/sheets/torgeternityActiveEffectConfig.js
@@ -38,6 +38,7 @@ export default class TorgActiveEffectConfig extends foundry.applications.sheets.
       partContext.transferToClass = this.document.system.transferOnOutcome ? '' : 'hidden';
       partContext.applyOutcomeClass = this.document.system.transferOnOutcome ? 'hidden' : '';
       partContext.transferOutcomeClass = this.document.system.applyOnOutcome ? 'hidden' : '';
+      partContext.showEmanation = (game.release.generation > 13);
     }
     return partContext;
   }